feat: card component
- Création des components
- Refacto du service
- Suppression des fichiers inutiles
