--- scripts/templates/photographer.js
@@ -1,1 +1,0 @@
-
